requirements.txt
torch>=2.0.0
torchvision
torchaudio
diffusers[torch]>=0.21.0
transformers>=4.35.0
accelerate
kaggle>=1.5.16
gradio>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib
seaborn
scikit-learn
scipy
mne>=1.5.0
Pillow
safetensors
datasets
hf_xet